+++ src/eval.c
@@ -33,6 +33,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "file_search.h"
 #include "garray.h"
 #include "move.h"
 #include "normal.h"
